feat: 동시성 문제 해결

+++ src/main/java/thegoods/server/data/dto/DataRequestDTO.java
@@ -40,7 +40,7 @@ public class DataRequestDTO {
         @NotNull
         Long dibsCount;
         @NotNull
-        Long salesCount;
+        Integer salesCount;
         @NotNull
         Long category;
         @Nullable
